wrote tests + fixed date issue

+++ src/deserialize.rs
@@ -8,7 +8,7 @@ pub struct Order {
     pub employee: String,
     pub client: String,
     pub description: String,
-    pub count: i64,
+    pub count: f64,
     pub ready: f64,
     pub leave: f64,
     pub start: f64,
@@ -65,10 +65,10 @@ pub fn deserialize_excel(file_path: &str) -> Result<Vec<Order>> {
                     count: row
                         .get(5)
                         .map(|x| match x {
-                            Data::String(x) => x.parse::<i64>().unwrap_or(0),
-                            Data::Float(x) => *x as i64,
-                            Data::Int(x) => *x,
-                            _ => 0,
+                            Data::String(x) => x.parse::<f64>().unwrap_or(0.0),
+                            Data::Float(x) => *x,
+                            Data::Int(x) => *x as f64,
+                            _ => 0.0,
                         })
                         .unwrap(),
                     ready: deserialize_date_cell(row.get(6), 0.0),
